+++ internal/connector/connector_test.go
@@ -2,18 +2,39 @@ package connector_test
 
 import (
 	"context"
+	"encoding/json"
 	"fmt"
 	"go.arcalot.io/assert"
 	"go.arcalot.io/log/v2"
 	"go.flow.arcalot.io/deployer"
 	"go.flow.arcalot.io/pluginsdk/atp"
 	"go.flow.arcalot.io/pluginsdk/schema"
+	"go.flow.arcalot.io/pythondeployer/internal/config"
 	pythondeployer "go.flow.arcalot.io/pythondeployer/pkg/factory"
-	"go.flow.arcalot.io/pythondeployer/tests"
+	"math/rand"
 	"os"
+	"os/exec"
 	"sync"
 	"testing"
 )
+
+type TestModule struct {
+	Location string
+	StepID   string
+	Input    map[string]any
+}
+
+func GetPythonPath() (string, error) {
+	python3Path, errPython3 := exec.LookPath("python3")
+	if errPython3 != nil {
+		pythonPath, errPython := exec.LookPath("python")
+		if errPython != nil {
+			return "", fmt.Errorf("error getting Python3 (%s) and python (%s)", errPython3, errPython)
+		}
+		return pythonPath, nil
+	}
+	return python3Path, nil
+}
 
 const examplePluginNickname string = "pythonuser"
 
@@ -32,7 +53,7 @@ func TestRunStepGit(t *testing.T) {
 		"name": examplePluginNickname,
 	}
 
-	connector, _ := tests.GetConnector(t, inOutConfigGitPullIfNotPresent, nil)
+	connector, _ := GetConnector(t, inOutConfigGitPullIfNotPresent, nil)
 	OutputID, OutputData, Error := RunStep(t, connector, moduleName, stepID, input)
 	assert.NoError(t, Error)
 	assert.Equals(t, OutputID, "success")
@@ -138,7 +159,7 @@ func TestDeployConcurrent_ConnectorsAndPluginsWithDifferentModules(t *testing.T)
 	unserializedConfig, err := deployerSchema.UnserializeType(serializedConfig)
 	assert.NoError(t, err)
 
-	pythonPath, err := tests.GetPythonPath()
+	pythonPath, err := GetPythonPath()
 	assert.NoError(t, err)
 	unserializedConfig.PythonPath = pythonPath
 
@@ -185,4 +206,50 @@ func TestDeployConcurrent_ConnectorsAndPluginsWithDifferentModules(t *testing.T)
 	t.Cleanup(func() {
 		assert.NoError(t, os.RemoveAll(rootDir))
 	})
+}
+
+func CreateWorkdir(t *testing.T) string {
+	workdir := fmt.Sprintf("/tmp/%s", RandString(10))
+	if _, err := os.Stat(workdir); !os.IsNotExist(err) {
+		err := os.RemoveAll(workdir)
+		assert.NoError(t, err)
+	}
+	err := os.Mkdir(workdir, os.ModePerm)
+	assert.NoError(t, err)
+	return workdir
+}
+
+func RandString(n int) string {
+	var chars = []rune("abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ")
+	b := make([]rune, n)
+	for i := range b {
+		b[i] = chars[rand.Intn(len(chars))]
+	}
+	return string(b)
+}
+
+func GetConnector(t *testing.T, configJSON string, workdir *string) (deployer.Connector, *config.Config) {
+	var serializedConfig any
+	if err := json.Unmarshal([]byte(configJSON), &serializedConfig); err != nil {
+		t.Fatal(err)
+	}
+	f := pythondeployer.NewFactory()
+	connectorSchema := f.ConfigurationSchema()
+	unserializedConfig, err := connectorSchema.UnserializeType(serializedConfig)
+	assert.NoError(t, err)
+	pythonPath, err := GetPythonPath()
+	assert.NoError(t, err)
+	unserializedConfig.PythonPath = pythonPath
+	// NOTE: randomizing Workdir to avoid parallel tests to
+	// remove python folders while other tests are running
+	// causing the test to fail
+	if workdir == nil {
+		unserializedConfig.WorkDir = CreateWorkdir(t)
+	} else {
+		unserializedConfig.WorkDir = *workdir
+	}
+
+	connector, err := f.Create(unserializedConfig, log.NewTestLogger(t))
+	assert.NoError(t, err)
+	return connector, unserializedConfig
 }
